Scanner_Container/requirements.txt
numpy==1.17.2
opencv_python==4.4.0.42
matplotlib==3.1.1
Flask==1.1.1
Pillow==7.2.0
